Improve hash-function usage
Two things: First, seeding the initial hash value with the depth number
directly does not really work. The different hash values are not all
that independent (at least given the implementation of boost::hash).
Second, we don't need to call the hash function once per depth-lookup
because the hash function generates 64-bits of hash, and we don't need
that many bits (the width of the CMS is just not that big). Thus, we can
call the hash function once and then use different parts of it for
different depth lookups.